Ejercicio actulizado con sanctum

+++ routes/api.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\API\AuthController;
 use App\Http\Controllers\NaveController;
 use App\Http\Controllers\PilotoController;
 use App\Http\Controllers\MantenimientoController;
+use App\Http\Controllers\UserController;
 use App\Models\Nave;
 use App\Models\Piloto;
 use Illuminate\Http\Request;
@@ -17,23 +18,35 @@ use Illuminate\Support\Facades\Route;
  * update: para actualizar un recurso existente
  * destroy: para eliminar un recurso
  */
+
+
+
+
 Route::middleware('auth:sanctum')->group(function () {
-    /**Endpoints naves */
-    Route::get('/naves', [NaveController::class, 'index']);
-    Route::get('/naves/{nave}', [NaveController::class, 'show']);
-    Route::post('/naves', [NaveController::class, 'store']);
-    Route::put('/naves/{nave}', [NaveController::class, 'update']);
-    Route::delete('/naves/{nave}', [NaveController::class, 'destroy']);
-    Route::post('/naves/asignarPiloto/{nave}', [NaveController::class, 'asignarPiloto']);
-    Route::put('/naves/desasignarPiloto/{nave}', [NaveController::class, 'desasignarPiloto']);
-    Route::get('/navesSinPiloto/', [NaveController::class, 'navesSinPiloto']);
+        /**Endpoints naves */
+    /* disponibles para todos*/
+    Route::get('/naves', [NaveController::class, 'index'])->middleware('midnavelistar');
+    Route::get('/naves/{nave}', [NaveController::class, 'show'])->middleware('midnavelistar');
+    Route::get('/navesSinPiloto/', [NaveController::class, 'navesSinPiloto'])->middleware('midnavelistar');
+    /**Solo accesible para admin */
+    Route::post('/naves', [NaveController::class, 'store'])->middleware('midadmin');
+    Route::put('/naves/{nave}', [NaveController::class, 'update'])->middleware('midadmin');
+    Route::delete('/naves/{nave}', [NaveController::class, 'destroy'])->middleware('midadmin');
+    Route::put('users/{user}/role', [UserController::class, 'update_role'])->middleware('midadmin');
+
+
+    /**Accesible para el gestor */
+    Route::post('/naves/asignarPiloto/{nave}', [NaveController::class, 'asignarPiloto'])->middleware('midpilotoasignar');
+    Route::put('/naves/desasignarPiloto/{nave}', [NaveController::class, 'desasignarPiloto'])->middleware('midpilotodesasignar');
+    
     /**Endpoints pilotos */
-    Route::get('/historicoPilotosAsignados', [PilotoController::class, 'listarHistoricoPilotosAsignados']);
-    Route::get('/pilotosAsignadosActualmente', [PilotoController::class, 'pilotosAsignadosActualmente']);
+    /**disponibles para todos */
+    Route::get('/historicoPilotosAsignados', [PilotoController::class, 'listarHistoricoPilotosAsignados'])->middleware('midpilotolistar');
+    Route::get('/pilotosAsignadosActualmente', [PilotoController::class, 'pilotosAsignadosActualmente'])->middleware('midpilotolistar');
     /**Endpoints mantenimientos */
-    Route::post('/naves/{nave}/mantenimientos ', [MantenimientoController::class, 'store']);
-    Route::get('/mantenimientos/{mantenimiento}', [MantenimientoController::class, 'show']);
-    Route::get('/mantenimientos/{inicio?}/{fin?}', [MantenimientoController::class, 'mantenimientosEntreFechas']);
+    Route::post('/naves/{nave}/mantenimientos ', [MantenimientoController::class, 'store'])->middleware('midmantenimientocrear');
+    Route::get('/mantenimientos/{mantenimiento}', [MantenimientoController::class, 'show'])->middleware('midmantenimientolistar');
+    Route::get('/mantenimientos/{inicio?}/{fin?}', [MantenimientoController::class, 'mantenimientosEntreFechas'])->middleware('midmantenimientolistar'); 
 });
 
 
@@ -44,3 +57,39 @@ Route::post('register', [AuthController::class, 'register']);
 Route::get('/nologin', function () {
     return response()->json(["success"=>false, "message" => "Unauthorised"],203);
 });
+
+// Route::middleware('auth:sanctum')->group(function () {
+//         /**Endpoints naves */
+//     /* disponibles para todos*/
+//     Route::get('/naves', [NaveController::class, 'index'])->middleware('ability:nave:listar');
+//     Route::get('/naves/{nave}', [NaveController::class, 'show'])->middleware('ability:nave:listar');
+//     Route::get('/navesSinPiloto/', [NaveController::class, 'navesSinPiloto']);
+//     /**Solo accesible para admin */
+//     Route::post('/naves', [NaveController::class, 'store'])->middleware('ability:*');
+//     Route::put('/naves/{nave}', [NaveController::class, 'update'])->middleware('ability:*');
+//     Route::delete('/naves/{nave}', [NaveController::class, 'destroy'])->middleware('ability:*');
+//     Route::put('users/{user}/role', [UserController::class, 'update_role'])->middleware('ability:*');
+
+
+//     /**Accesible para el gestor */
+//     Route::post('/naves/asignarPiloto/{nave}', [NaveController::class, 'asignarPiloto'])->middleware('ability:piloto:asignar');
+//     Route::put('/naves/desasignarPiloto/{nave}', [NaveController::class, 'desasignarPiloto'])->middleware('ability:piloto:desasignar');
+    
+//     /**Endpoints pilotos */
+//     /**disponibles para todos */
+//     Route::get('/historicoPilotosAsignados', [PilotoController::class, 'listarHistoricoPilotosAsignados'])->middleware('ability:piloto:listar');
+//     Route::get('/pilotosAsignadosActualmente', [PilotoController::class, 'pilotosAsignadosActualmente'])->middleware('ability:piloto:listar');
+//     /**Endpoints mantenimientos */
+//     Route::post('/naves/{nave}/mantenimientos ', [MantenimientoController::class, 'store'])->middleware('ability: mantenimiento:crear');
+//     Route::get('/mantenimientos/{mantenimiento}', [MantenimientoController::class, 'show'])->middleware('ability: mantenimiento:listar');
+//     Route::get('/mantenimientos/{inicio?}/{fin?}', [MantenimientoController::class, 'mantenimientosEntreFechas'])->middleware('ability:mantenimiento:listar'); 
+// });
+
+
+// Route::post('login', [AuthController::class, 'login']);
+// Route::post('logout', [AuthController::class, 'logout']);
+// Route::post('register', [AuthController::class, 'register']);
+
+// Route::get('/nologin', function () {
+//     return response()->json(["success"=>false, "message" => "Unauthorised"],203);
+// });
